+++ exonum/src/storage/proof_list_index/mod.rs
@@ -16,12 +16,15 @@
 
 pub use self::proof::{ListProof, ListProofError};
 
-use std::cell::Cell;
-use std::marker::PhantomData;
+use std::{cell::Cell, marker::PhantomData};
 
 use self::key::ProofListKey;
-use super::indexes_metadata::IndexType;
-use super::{BaseIndex, BaseIndexIter, Fork, Snapshot, StorageKey, StorageValue};
+use super::{base_index::{BaseIndex, BaseIndexIter},
+            indexes_metadata::IndexType,
+            Fork,
+            Snapshot,
+            StorageKey,
+            StorageValue};
 use crypto::{hash, Hash, HashStream};
 
 mod key;