+++ services/time/examples/simple_service.rs
@@ -24,14 +24,14 @@ extern crate serde;
 extern crate serde_json;
 
 use chrono::{DateTime, Duration, TimeZone, Utc};
-use exonum::blockchain::{ExecutionResult, Service, Transaction, TransactionSet};
-use exonum::crypto::{gen_keypair, Hash, PublicKey};
-use exonum::encoding;
-use exonum::helpers::Height;
-use exonum::messages::{Message, RawTransaction};
-use exonum::storage::{Fork, ProofMapIndex, Snapshot};
+use exonum::{blockchain::{ExecutionResult, Service, Transaction, TransactionSet},
+             crypto::{gen_keypair, Hash, PublicKey},
+             encoding,
+             helpers::Height,
+             messages::{Message, RawTransaction},
+             storage::{Fork, ProofMapIndex, Snapshot}};
 use exonum_testkit::TestKitBuilder;
-use exonum_time::{TimeService, schema::TimeSchema, time_provider::MockTimeProvider};
+use exonum_time::{schema::TimeSchema, time_provider::MockTimeProvider, TimeService};
 
 /// Marker service id.
 const SERVICE_ID: u16 = 128;