+++ CHANGELOG.md
@@ -5,14 +5,41 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+### Breaking changes
+
+#### exonum
+
+- `storage::base_index` module has become private along with `BaseIndex` and
+  `BaseIndexIter` types. (#723)
+
+### New features
+
+#### exonum
+
+- `exonum::crypto::x25519` module to convert from Ed25519 keys to X25519 keys
+  has been introduced.
+
+### Internal improvements
+
+#### exonum
+
+- `handle_consensus` now does not write warning for message from previous
+  height. (#729)
+
+### Bug fixes
+
+#### exonum
+
+- Fixed bug with incorrect peer status for turned off node. (#730)
+
 ## 0.8 - 2018-05-31
 
 ### Breaking changes
 
+#### exonum
+
 - `handle_commit` method in `Service` trait  has been renamed to
   `after_commit`. (#715)
-
-#### exonum
 
 - `TimeoutAdjusterConfig` has been removed along with different timeout
   adjusters. Current behavior is similar to the `Dynamic` timeout adjuster and
@@ -33,10 +60,18 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   [Noise Protocol](https://noiseprotocol.org/). Nodes compiled with old
   version will not connect to the new ones. Therefore you need to
   update all node instances for the network to work. (#678)
+
 - `storage::Error` constructor has been made private. (#689)
 
 - `ConsensusConfig::validate_configuration` method has been renamed to the
   `warn_if_nonoptimal`. (#690)
+
+#### exonum-time
+
+- The service has been refactored and the following public structs has been
+  moved to separate modules: `TimeSchema` to `exonum_time::schema`,
+  `TimeProvider` and `MockTimeProvider` to `exonum_time::time_provider`,
+  `ValidatorTime` to `exonum_time::api`. (#604)
 
 ### New features
 
@@ -111,10 +146,6 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   reduce boilerplate. (#639)
 
 - Metrics are now using `chrono::DateTime<Utc>` instead of `SystemTime`. (#620)
-
-#### exonum-time
-
-- Split service components to separate modules. (#604)
 
 #### exonum-configuration